Refactor Car package structure; Implement User package
The diff shows a significant refactoring of the car-controlled features in the application with the introduction of a new user-controlled feature. The rename changes are due to separating different entities (Car and User) into their packages. This refactoring will help in better organization and readability of the codebase.

In the car-controlled feature, there is a change in the method addCar() to taking a NewCar object instead of a Car object. It reflects the idea of creating a new Car entity with id property being generated in the service layer.

We also introduced a user-controlled feature that includes the implementation of UserController, UserService, UserRepository, and User model. This addition facilitates the management of User entities in our application.